+++ example/lib/main.dart
@@ -1,8 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:rxflare/rxflare.dart';
 
+import 'auto_dispose/data_dispose.dart';
 import 'feature/features/home/view/freature_home_page.dart';
 import 'router_demo.dart';
+import 'auto_dispose/auto_dispose.dart';
 
 final isDarkMode = false.obs;
 
@@ -42,7 +44,7 @@ class _HomePageState extends State<HomePage> with SingleTickerProviderStateMixin
   @override
   void initState() {
     super.initState();
-    _tabController = TabController(length: 8, vsync: this);
+    _tabController = TabController(length: 10, vsync: this);
   }
 
   @override
@@ -72,12 +74,16 @@ class _HomePageState extends State<HomePage> with SingleTickerProviderStateMixin
             Tab(text: '性能测试2'),
             Tab(text: '路由'),
             Tab(text: 'RxFuture用法'),
+            Tab(text: '简单生命周期'),
+            Tab(text: '复杂生命周期'),
           ],
         ),
       ),
       body: TabBarView(
         controller: _tabController,
-        children: [BasicAutoTrackDemo(), MapListPreciseDemo(), ComputedDemo(), EventBusDemo(), PerformanceTestDemo(), PerformanceTestDemoA(), RouterDemo(),FreatureHomePage()],
+        children: [BasicAutoTrackDemo(), MapListPreciseDemo(), ComputedDemo(), EventBusDemo(), PerformanceTestDemo(), 
+        PerformanceTestDemoA(), RouterDemo(),FreatureHomePage(),
+        AutoDisposePage(),DataDispose()],
       ),
     );
   }